backend/requirements.txt
# ===============================
# PyTorch CPU (บังคับ CPU ล้วน)
# ===============================
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.1+cpu
torchvision==0.17.1+cpu
torchaudio==2.2.1+cpu

# ===============================
# Computer Vision
# ===============================
ultralytics==8.4.14
opencv-python-headless==4.10.0.84
pillow==10.3.0

# ===============================
# Backend & API
# ===============================
fastapi==0.111.0
uvicorn[standard]==0.30.1
gunicorn==22.0.0
python-dotenv==1.0.1
python-multipart==0.0.9

# ===============================
# Database
# ===============================
SQLAlchemy==2.0.30

# (ถ้าใช้ async engine ให้เปิดบรรทัดนี้)
# asyncpg==0.29.0

# ===============================
# Supabase
# ===============================
supabase==2.5.0
httpx==0.27.0

# ===============================
# Data Processing (stable range)
# ===============================
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.4.0,<2.0.0

# ===============================
# Security & Auth
# ===============================
pydantic>=2.0.0,<3.0.0
cryptography>=42.0.0
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
Authlib==1.3.1
bcrypt==4.1.3
passlib[bcrypt]==1.7.4
requests==2.32.3
PyYAML==6.0.1
itsdangerous
psycopg2-binary==2.9.9
